Docs: Moved all pages from `pages` to `docs` - Replaced useRouter from next/router to next/navigation - Updated modal logic to use useRouters basic API - Removed next-seo as Next13 provides that out of the box
