+++ docs/package.json
@@ -20,7 +20,6 @@
     "@vercel/analytics": "^0.1.8",
     "cookie": "^0.5.0",
     "next": "^13.4.1",
-    "next-seo": "^5.4.0",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "shiki": "^0.14.2",
