+++ docs/src/app/thank-you/page.tsx
@@ -1,5 +1,5 @@
-import Navigation from '../components/navigation';
-import Footer from '../components/footer';
+import Navigation from '../../components/navigation';
+import Footer from '../../components/footer';
 
 export default function Index() {
   return (
