+++ docs/src/components/call-to-action.tsx
@@ -1,3 +1,5 @@
+'use client';
+
 import { useState } from 'react';
 import { ArrowTopRightOnSquareIcon } from '@heroicons/react/24/outline';
