+++ docs/src/hooks/useFormSubmission.ts
@@ -1,5 +1,5 @@
 import { useState } from 'react';
-import { useRouter } from 'next/router';
+import { useRouter } from 'next/navigation';
 
 const AIRTABLE_TOKEN = process.env.NEXT_PUBLIC_AIRTABLE_TOKEN;
 const BASE_ID = process.env.NEXT_PUBLIC_AIRTABLE_BASE_ID;
